Update README.md
The challenge project it has two folder which are ui_challenge and coding_challenge,a ui_challenge folder it building an interactive photo gallery by using HTML,CSS and javascript,and coding_challenge it has two files which are arrayMap.js this file is about an array of integers and a target sum to determine if there exists a contiguous subarray within the array that sums up to the target, it returns true if such a subarray exists,otherwise return false.and other file is stringTransform.js this file is about a string ,if the length of the string is divisible by 3,reverse the entire string,if the length of the string is divisible by 5,replace each character with its ASCII code, and if the length of the string is divisible by both 3 and 5(i.e divisible by 15), both files by using javascript. 
The process of set up this project firstly,you copy the clone link in my github after that you open your text editor and then  you go to your text editor's terminal and write a command for paste the git clone link,then after the installation finish you will be able to see the project.
To run this project first of all it has two folder which are ui_challenge this one you can go to index.js file and click on right click and select open with live server it will take you to the browser and see the output and othe folder is coding_challenge this one you can run it by write the command in the terminal "node filename.js".